+++ src/MIDI.jl
@@ -10,5 +10,6 @@ include("midifile.jl")
 include("constants.jl")
 include("variablelength.jl")
 include("convert.jl")
+include("findevents.jl")
 
 end