+++ src/miditrack.jl
@@ -201,7 +201,7 @@ function addevents!(track::MIDITrack, times::AbstractArray{Int}, events)
 
     # get a permutation that gives temporal order
     if issorted(times)
-        perm = collect(1:length(times))
+        perm = 1:length(times)
     else
         perm = sortperm(times)
     end
@@ -210,7 +210,9 @@ function addevents!(track::MIDITrack, times::AbstractArray{Int}, events)
     eventindex = 0
     eventtime = 0
     for i = 1:length(times)
-        eventindex, eventtime = addevent_hint!(track,times[perm[i]], events[perm[i]], eventindex, eventtime)
+        eventindex, eventtime = addevent_hint!(
+            track, times[perm[i]], events[perm[i]], eventindex, eventtime
+        )
     end
 end
 
@@ -301,4 +303,25 @@ function programchange(track::MIDITrack, time::Integer, channel::UInt8, program:
     "and be kind enough to report whether it worked!"
     program -= 1
     addevent!(track, time, MIDIEvent(0, PROGRAMCHANGE | channel, UInt8[program]))
+end
+
+
+"""
+    get_abs_pos(track::MIDITrack, idxs)
+Return the absolute positions (since track start) of the events given by the
+indices `idxs` of `track.events`.
+"""
+function get_abs_pos(idxs, track::MIDITrack)
+    abspos = Int[]
+    evtime = 0
+    j = 1; L = length(idxs)
+    for (i, event) in enumerate(track.events)
+        evtime += event.dT
+        if i == idxs[j]
+            push!(abspos, evtime)
+            j += 1
+            j == L + 1 && return abspos
+        end
+    end
+    return abspos
 end