+++ CHANGELOG.md
@@ -20,6 +20,8 @@ work in progress changes are contained in this section.
 ## Improvements
 * `BPM` is now faster as it iterates over events.
 * Pretty printing for `MIDIFile`.
+* Added functions `textevent` and `findtextevents` that create and find
+  text related meta events (lyrics, text, markers).
 
 # v0.6.2